First commit. Source and some simple example graphics.
Creates a world from a height map, textures it from a texture map, and
allows the player to move and look around.

In theory, most of this should work.However it appears to be
plagued with graphical glitches, especially when tested on my phone. I
think multi-touch works, but with the glitches it was hard to be sure.

The glitches got worse when the ground texture atlas was made smaller,
and got comparatively better when the image was made larger - so
maybe image scaling has something to do with it.

Height map and texture map interpreting code is basic - enough so that
multiple textures could be tested.